+++ src/components/Expenses/ExpenseItem.js
@@ -1,6 +1,6 @@
 import "./ExpenseItem.css";
 import ExpenseDate from "./ExpenseDate";
-import Card from './Card';
+import Card from '..UI/Card';
 
 function ExpenseItem(props) {
